refactor: module info
