+++ ui/src/components/modules/index.ts
@@ -1,0 +1,5 @@
+export { ModuleDetailsCard } from './ModuleDetailsCard'
+export { ExportsCard } from './ExportsCard'
+export { ImportsCard } from './ImportsCard'
+export { PluginCard } from './PluginCard'
+export { WasiSupportCard } from './WasiSupportCard'
